oops float and double on firbird, thanks Jose

+++ gluon/dal.py
@@ -1302,7 +1302,7 @@ class FireBirdAdapter(BaseAdapter):
         'blob': 'BLOB SUB_TYPE 0',
         'upload': 'VARCHAR(%(length)s)',
         'integer': 'INTEGER',
-        'double': 'FLOAT',
+        'double': 'DOUBLE',
         'decimal': 'DECIMAL(%(precision)s,%(scale)s)',
         'date': 'DATE',
         'time': 'TIME',
